Home_task_1 add button
