+++ Elena_Bokhan/webpack.config.js
@@ -21,6 +21,7 @@ module.exports = {
 	},
 	plugins: [
 		new HTMLWebpackPlugin({template: path.resolve(__dirname, "src","index.html"),})
-	]
+	],
+	resolve: {extensions: ['.jsx','.js']}
 	
 }
